refactored fs interactions

+++ fs.go
@@ -5,7 +5,11 @@ import (
 	"log"
 	"os"
 	"path/filepath"
+	"slices"
 	"strings"
+	"time"
+
+	"github.com/urfave/cli/v3"
 )
 
 type DirEntry struct {
@@ -15,38 +19,73 @@ type DirEntry struct {
 	DeletionInProgress bool
 }
 
-func Traverse(dir string, searchDirs, ignoreDirs []string, searchAll bool, ch chan DirEntry) {
-	defer close(ch)
+type FsHandler struct {
+	Entries             chan DirEntry
+	directoriesToFind   []string
+	directoriesToIgnore []string
+	searchAll           bool
+	isDryRun            bool
+}
+
+func NewFsHandler(c *cli.Context) *FsHandler {
+	return &FsHandler{
+		Entries:             make(chan DirEntry),
+		searchAll:           c.Bool("all"),
+		directoriesToFind:   c.Args().Slice(),
+		directoriesToIgnore: c.StringSlice("exclude"),
+		isDryRun:            c.Bool("dry-run"),
+	}
+}
+
+func (h *FsHandler) Traverse() {
+	defer close(h.Entries)
+
+	if h.isDryRun {
+		for _, e := range []DirEntry{
+			{Path: "test/asd", Size: 2 * gb},
+			{Path: "test/asd2", Size: 500 * mb},
+			{Path: "test/asd3", Size: 10 * kb},
+		} {
+			time.Sleep(500 * time.Millisecond)
+			h.Entries <- e
+		}
+		return
+	}
 
 	// Scan the file tree starting from current working directory
 	// Don't return err because it stops the walk
 	// Instead just log it and skip the directory
-	filepath.WalkDir(dir, func(path string, d fs.DirEntry, err error) error {
+	filepath.WalkDir("", func(path string, d fs.DirEntry, err error) error {
 		if err != nil {
 			log.Println(err)
 			return filepath.SkipDir
 		}
 		if d.IsDir() {
-			if contains(searchDirs, d.Name()) {
+			if slices.Contains(h.directoriesToFind, d.Name()) {
 				size, err := getDirectorySize(path)
 				if err != nil {
 					log.Println(err)
 					return filepath.SkipDir
 				}
-				ch <- DirEntry{Path: path, Size: size}
+				h.Entries <- DirEntry{Path: path, Size: size}
 				return filepath.SkipDir
 			}
-			if contains(ignoreDirs, d.Name()) ||
-				(strings.HasPrefix(d.Name(), ".") && !searchAll) {
+			if slices.Contains(h.directoriesToIgnore, d.Name()) ||
+				(strings.HasPrefix(d.Name(), ".") && !h.searchAll) {
 				log.Println("Skipping ", path)
 				return filepath.SkipDir
 			}
 		}
+
 		return err
 	})
 }
 
-func Delete(path string) error {
+func (h *FsHandler) Delete(path string) error {
+	if h.isDryRun {
+		time.Sleep(500 * time.Millisecond)
+		return nil
+	}
 	return os.RemoveAll(path)
 }
 
@@ -70,13 +109,4 @@ func getDirectorySize(path string) (int64, error) {
 		return 0, err
 	}
 	return size, nil
-}
-
-func contains(s []string, e string) bool {
-	for _, a := range s {
-		if a == e {
-			return true
-		}
-	}
-	return false
 }
